Fix #2 and fix #3
The former requires some way to resolve a type reference of a record
field to its full type in the context of the macro.  This proved
somewhat too involved to solve inside Jinja2 macros and so a number of
Python helper functions are now given to Jinja2.  Some are generic and
some are specific to a target language and made available as
"model.<lang>.*" eg "model.cpp.field_default()".

In any case, this also now requires two changes in an "omodel" and an
"oschema".  In addition to the types a schema defines, a schema must
also provide all types that are referenced by those it defines.  The
model provides all as "all_types" and then applies a path filter to
define a subset "types" that are then used as subjects of codegen.

As the code grew for this apparently simple fix, the template.py has
been moved to multiple modules under moo/templates/.

There was no particualr change needed to provide for nested structs
and this new default value mechanism works for it as well given the
default value is itself valid for the schema of the field type.