Fix #9 and break environment variable handling.
Now env. var. are not honored outside of __main__.py.  For the CLI we
honor MOO_LOAD_PATH instead of the previous and more confusingly named
MOO_MODEL_PATH.  The MOO_TEMPLATE_PATH is left untouched.

Application developers may set moo.io.default_load_path to avoid
explicit passing of paths to moo.io.load() or functions that forward
to it.

The otypes.org doc mentioned this issue and is updated to reflect the
changes.